+++ src/zksync/mod.rs
@@ -2,4 +2,5 @@ pub mod artifact_output;
 pub mod artifacts;
 pub mod cache;
 pub mod compile;
+pub mod compilers;
 pub mod config;
